Feat: 이벤트 소싱 기반 거래소 매칭 엔진 (#4)
* feat: add market data service

* feat: add LMAX Disruptor

+++ src/main/java/com/project/snaptrade/engine/Dto/OrderRequestDto.java
@@ -14,6 +14,6 @@ public class OrderRequestDto {
     private OrderSide side;
     private OrderType orderType;
     private TimeInForce timeInForce;
-    private BigDecimal price;
-    private BigDecimal quantity;
+    private Long price;
+    private Long quantity;
 }
